+++ src/apps/seelen_task_switcher_rename_later/index.tsx
@@ -6,14 +6,12 @@ import { store } from './modules/shared/store/infra';
 import { App } from './app';
 
 import { getRootContainer } from '../shared';
-import { wrapConsole } from '../shared/ConsoleWrapper';
 
 import '../shared/styles/colors.css';
 import '../shared/styles/reset.css';
 import './styles/global.css';
 
 async function Main() {
-  wrapConsole();
   const container = getRootContainer();
   createRoot(container).render(
     <Provider store={store}>