+++ src/apps/seelenweg/index.tsx
@@ -10,7 +10,6 @@ import { loadStore, registerStoreEvents, store } from './modules/shared/store/in
 import { App } from './app';
 
 import { getRootContainer } from '../shared';
-import { wrapConsole } from '../shared/ConsoleWrapper';
 import i18n, { loadTranslations } from './i18n';
 
 import '../shared/styles/colors.css';
@@ -18,7 +17,8 @@ import './styles/variables.css';
 import '../shared/styles/reset.css';
 import './styles/global.css';
 
-wrapConsole();
+throw new Error('Not implemented');
+
 await declareDocumentAsLayeredHitbox();
 await loadStore();
 await registerStoreEvents();