+++ src/apps/settings/index.tsx
@@ -8,7 +8,6 @@ import { LoadSettingsToStore, registerStoreEvents, store } from './modules/share
 import { App } from './app';
 
 import { getRootContainer } from '../shared';
-import { wrapConsole } from '../shared/ConsoleWrapper';
 import i18n, { loadTranslations } from './i18n';
 
 import '../shared/styles/colors.css';
@@ -17,7 +16,6 @@ import '../shared/styles/reset.css';
 import './styles/global.css';
 
 (async function main() {
-  wrapConsole();
   getCurrentWebviewWindow().show();
   const container = getRootContainer();