+++ src/background/state/application/apps_config.rs
@@ -7,22 +7,15 @@ use super::FullState;
 impl FullState {
     pub fn get_app_config_by_window(&self, hwnd: HWND) -> Option<&AppConfig> {
         let window = Window::from(hwnd);
-        if let Ok(path) = window.process().program_path() {
-            let title = window.title();
-            let class = window.class();
-            let exe = path
-                .file_name()
-                .unwrap_or_default()
-                .to_string_lossy()
-                .to_string();
-            let path = path.to_string_lossy().to_string();
 
-            for app in self.settings_by_app.iter() {
-                if app.identifier.validate(&title, &class, &exe, &path) {
-                    return Option::from(app);
-                }
-            }
-        }
-        None
+        let path = window.process().program_path().ok()?;
+        let title = window.title();
+        let class = window.class();
+        let exe = path.file_name()?.to_string_lossy().to_string();
+        let path = path.to_string_lossy().to_string();
+
+        self.settings_by_app
+            .iter()
+            .find(|&config| config.identifier.validate(&title, &class, &exe, &path))
     }
 }