+++ src/background/modules/notifications/infrastructure.rs
@@ -9,7 +9,9 @@ use crate::{
 use super::application::NOTIFICATION_MANAGER;
 
 pub fn register_notification_events() {
-    log_error!(trace_lock!(NOTIFICATION_MANAGER).initialize());
+    std::thread::spawn(|| {
+        log_error!(trace_lock!(NOTIFICATION_MANAGER).initialize());
+    });
     NotificationManager::subscribe(|_event| {
         log_error!(get_app_handle().emit(
             SeelenEvent::Notifications,