+++ changelog.md
@@ -4,6 +4,9 @@
 ### enhancements
 - wrap webview console as soon as posible to avoid missing errors on logs.
 
+### fix
+- panic on media module when loading initial devices.
+
 ## [2.2.2]
 ### enhancements
 - add option to disable dynamic colors on toolbar.